+++ server/src/test/java/org/eclipse/openvsx/TestDatabaseConfig.java
@@ -1,32 +1,30 @@
-/********************************************************************************
- * Copyright (c) 2020 TypeFox and others
+/******************************************************************************
+ * Copyright (c) 2026 Contributors to the Eclipse Foundation.
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
  *
  * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
+ * terms of the Eclipse Public License 2.0 which is available at
+ * https://www.eclipse.org/legal/epl-2.0.
  *
  * SPDX-License-Identifier: EPL-2.0
- ********************************************************************************/
+ *****************************************************************************/
 package org.eclipse.openvsx;
 
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.testcontainers.elasticsearch.ElasticsearchContainer;
-import org.testcontainers.utility.DockerImageName;
+import org.testcontainers.postgresql.PostgreSQLContainer;
 
 @Configuration
-@Profile("test")
+@Profile("test_db")
 public class TestDatabaseConfig {
 
     @Bean
     @ServiceConnection
-    ElasticsearchContainer elasticsearchContainer() {
-        return new ElasticsearchContainer(
-                DockerImageName.parse("elasticsearch:8.17.3")
-                        .asCompatibleSubstituteFor("docker.elastic.co/elasticsearch/elasticsearch"))
-                .withEnv("discovery.type", "single-node")
-                .withEnv("xpack.security.enabled", "false");
+    PostgreSQLContainer postgreSQLContainer() {
+        return new PostgreSQLContainer("postgres:16.2");
     }
 }
