+++ server/src/test/java/org/eclipse/openvsx/ratelimit/RateLimitIntegrationTest.java
@@ -24,11 +24,8 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.resttestclient.TestRestTemplate;
 import org.springframework.boot.resttestclient.autoconfigure.AutoConfigureTestRestTemplate;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.postgresql.PostgreSQLContainer;
 import redis.clients.jedis.RedisClusterClient;
 
 import java.net.URI;
@@ -43,12 +40,8 @@ import static org.mockito.ArgumentMatchers.any;
         "ovsx.elasticsearch.enabled=false"
 })
 @AutoConfigureTestRestTemplate
-@Testcontainers
+@ActiveProfiles("test_db")
 class RateLimitIntegrationTest {
-
-    @Container
-    @ServiceConnection
-    static PostgreSQLContainer postgres = new PostgreSQLContainer("postgres:16.2");
 
     @LocalServerPort
     int port;
